refactor: store에 빈 값이 들어가는 경우를 방지, store를 분리하지 않고 하나로 사용
