+++ index.html
@@ -17,14 +17,6 @@
   </h1>
   <section>
     <div id="user-list">
-      <button class="ripple active">makerjun</button>
-      <button class="ripple">eastjun</button>
-      <button class="ripple user-create-button" data-action="createUser">
-        + 유저 생성
-      </button>
-      <button class="ripple user-delete-button" data-action="deleteUser">
-        삭제 -
-      </button>
     </div>
   </section>
